requirements.txt
Django==1.11.5
django-crispy-forms==1.6.1
django-pagedown==0.1.3
djangorestframework==3.6.4
Markdown==2.6.9
mysqlclient==1.3.12
pytz==2017.2
c